fixed timing issue in injected events
this is a fix for counting UniqueClients, which depends on the times
of the injected events being within a window around the current
timestamp. this causes issues when injecting events from the past
(like those genereated in shadow). this change rewrites the event
times to the current time to handle this situation.