The TS produces an "outcome" file containing additional context
The outcome file contains:
* Tally (existing "tallies" file)
* Context
  * DataCollector: UID, Status and Config
  * ShareKeeper: UID, Status and Config
  * TallyServer: Status and Config
  * Time: ClockPadding, Period, Start, Stop
(A few of the status and config items are excluded.)

Clients send their config information along with the counts at the end
of the run.

Including all status and config information makes it far less likely
we finish a run, and then work out we needed more context. It also
makes it easier to add more context in future.