Add event_period, replacing the default TS refresh_loop of 1 minute
There must be at least two event_periods and checkin_periods in each
collect_period.

The checkin_period should be less than the event_period, otherwise
the TS might not react to client status updates in a timely fashion.
(This is far less of an issue if the collect_period is large, and
the event_period and checkin_period relatively small.)

Closes #7