Add basic sanity checks for noise weight and DC threshold
The TS sends the noise weights and DC threshold to all DCs and SKs.

All parties check that:
* the DC threshold is between 1 and 1 million
* the noise weights are between 0.0 and the tallied counter maximum
* the total noise weight is between 0.0 and the tallied counter maximum

These checks can be further expanded in #27.

DCs fail if there is no noise weight for their fingerprint.
(DCs only discover their fingerprint when they connect to the control port,
so the round starts, then failed DCs stop and delete their counters. When the
round ends, failed DCs do not return any counters.)

Always send DC configs to TS, even when counts fail or are not requested.
(This assists in debugging.)