PrivCount can count parameters for traffic modeling (#198)
* PrivCount can count parameters for traffic modeling

PrivCount takes in an HMM traffic model, and counts some items
needed to understand how well the data fits those models. The
output of a counting round, i.e., the tallies, are used to update
the traffic model for future rounds. Noise for the traffic model
file is specified in a special noise file containing template
labels that PrivCount expands to assign noise to all counters.

Nick Hopper contributed to the traffic model viterbi algorithm
and to the traffic model update function (thanks!).

closes #19