Compare floats using their string values in the unit tests
Renaming counters changed the order in which they are calculated, which
made some floating point values become close-but-not-equal.

This only seems to affect noise calculation at the moment, but could
potentially affect any part of the protocol that compares floating point
values.

A quick way of fixing this is to compare string values rather than float
values.

Related to #69, where we will solve this reliably and permanently.

Closes #260