Make clients retry a connection every checkin interval
Before this change, clients only tried to connect once after one checkin
interval had elapsed. Because we use a ReconnectingClient, this worked, but
only when the outage was short, and the exponential backoff remained low.

Also, allow an extra period for the client to checkin before warning and
before declaring it dead. Allow for the RTT plus a small amount of variance.

Prevents spurious warnings in networks with low checkin times.

Fixes #274, related to #211.

Also contains some partial fixes to #47:
* always stop the reactor on an uncaught exception,
* always stop scheduled tasks on an error,
  (previously, the deferred was cancelled, but the task was still scheduled),
* log a brief stack trace at warning level on exceptions in Deferred,
  (and a full stack trace with variables at debug level).

Also fix a comment typo.