Make the default checkin and event periods 10 minutes
This reduces load on the tally server, particularly when a large number
of nodes are participating.

It also increases the default allowed amount of downtime for a node to
just over an hour. (If a node is down for an hour, it's probably missing
from the consensus, so we do want to declare it dead.)

Fixes #378.