Allow matching against multiple domain lists
The first domain list is used for the ExitDomain*MatchStream Ratio,
LifeTime, and Histogram counters. The corresponding
ExitDomainNo*MatchStream* counters are used when there is no match in the
first list.

If there is more than one domain list configured, each domain list
corresponds to a bin in the ExitDomain*MatchStream*CountList counters.
If multiple lists match, each corresponding bin is incremented.
The final bin is incremented for domains that don't match any list.

Removes 24 ExitDomainNo*Match*Count counters, because we use the final
bin in the ExitDomain*MatchStream*CountList counters for "no match".

Adds the following domain lists for testing:
  - domain-top-1k.txt  : Alexa Top 1000, November 2017
  - domain-tld.txt     : IANA Top-Level Domains, November 2017
  - domain-cc.txt      : IANA Country Codes, November 2017
  - domain-local.txt   : IANA reserved domains and common internal domains
  - domain-example.txt : IANA example domains
  - domain-arpa.txt    : IANA arpa infrastructure domain
  - domain-onion.txt   : Tor's onion special use domain
  - domain-i2p.txt     : I2P's i2p special use domain

Closes #179.