Implement a more efficient tree structure for domain matching
This makes domain suffix matching about 20% faster, and saves about 10% RAM,
at the cost of about 30s preparation at the start of the round on the tally
server (for the alexa top 1 million).

We can either repeat this 30s preparation on the Data Collectors, or send
the processed list for 15MB (for the top 1 million). This commit sends the
processed list, so it increases the maximum line length to 200MB.

Part of #469.