Rewrote and fixed tmodel update code to produce valid new models
Rewrote the code that transforms a configured traffic model using
the tally results for a round, so that the updated model that it
produces is valid according to the traffic model checker code.

Added sample traffic model tallies and updated the traffic model
test case to test that a tranformed model is valid.